backend/app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
api-jwt==1.2.2
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
dependency-injector==4.41.0
ecdsa==0.18.0
fastapi==0.109.2
fastapi-jwt==0.2.0
fastapi-pagination==0.12.15
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
packaging==23.2
# Have problem with bcrypt higher 4.0
# passlib==1.7.4
# Fork which fix problem with bcrypt
passlib-fork==1.7.4.post2
Pillow==10.2.0
prisma==0.12.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.10.3
reportlab==4.1.0
requests==2.31.0
rsa==4.9
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tokenizers==0.15.2
tomlkit==0.12.3
torch==2.2.0
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
funcy==2.0
python-dateutil==2.9.0.post0
